requirements.txt
neuralhydrology
hydra-core
hydra_colorlog
earthengine-api
geemap
ipykernel